+++ packages/vsce/src/webviews/common/RefreshButton.tsx
@@ -1,0 +1,19 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+export const RefreshButton = ({ onClick, tabIndex }: { onClick: () => void; tabIndex?: number; }) => {
+  return <button
+    id="refresh-icon"
+    className="codicon codicon-refresh rotate-45 cursor-pointer font-bold bg-transparent border-0 p-0"
+    onClick={onClick}
+    tabIndex={tabIndex}
+  />;
+};